requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Google AI
google-generativeai==0.3.2

# Databases
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
motor==3.3.2
boto3==1.34.34

# Market Data
yfinance==0.2.35
pandas==2.1.4
ta==0.11.0

# HTTP & Async
aiohttp==3.9.1
httpx==0.26.0

# Auth
python-jose[cryptography]==3.3.0

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
loguru==0.7.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
mypy==1.8.0
